Split vtsearch/media/base.py into four focused modules
base.py was 1420 lines mixing four unrelated plugin ABCs. Split into:
- base.py (~380 lines): MediaType, MediaResponse, DemoDataset, ProgressCallback
- embedder.py (~350 lines): MediaEmbedder ABC + model-loading helpers
- processors.py (~200 lines): Processor, Detector, Localizer, Extractor
- clipper.py (~155 lines): MediaClipper ABC

All direct imports updated across source, tests, and docs.
The __init__.py re-exports are unchanged so `from vtsearch.media import X`
continues to work.

This breaks backwards compatibility for direct `from vtsearch.media.base`
imports of moved symbols.
