+++ tests/test_preload_progress.py
@@ -17,7 +17,7 @@ from unittest.mock import MagicMock, patch
 import torch
 import torch.nn as nn
 
-from vtsearch.media.base import intercept_weight_loading_progress, load_pretrained_local_first, timed_progress
+from vtsearch.media.embedder import intercept_weight_loading_progress, load_pretrained_local_first, timed_progress
 from vtsearch.models.loader import _make_console_progress, preload_autoload_media_types
