+++ vtsearch/media/video/clipper.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import math
 from typing import Any
 
-from vtsearch.media.base import MediaClipper
+from vtsearch.media.clipper import MediaClipper
 
 
 class VideoDefaultClipper(MediaClipper):
